nvidia: add nvidia attester back to all-attesters
Make the NVAT dependency conditional on the platform (only require it on
x86). This allows us to add the nvidia attester back to `all-attesters`
without breaking too many things.
