+++ attestation-agent/attester/Cargo.toml
@@ -23,7 +23,6 @@ crypto.path = "../deps/crypto"
 hex.workspace = true
 iocuddle = { version = "0.2.0", optional = true }
 kbs-types.workspace = true
-nv-attestation-sdk = { git = "https://github.com/NVIDIA/attestation-sdk", tag = "2026.03.02", optional = true }
 occlum_dcap = { git = "https://github.com/occlum/occlum", tag = "v0.29.7", optional = true }
 pem-rfc7468 = { version = "1", optional = true, features = ["alloc"] }
 pv = { version = "0.12.0", package = "s390_pv", optional = true }
@@ -60,6 +59,9 @@ zerocopy = { version = "0.8.54", optional = true }
 tokio.workspace = true
 rstest.workspace = true
 
+[target.'cfg(target_arch = "x86_64")'.dependencies]
+nv-attestation-sdk = { git = "https://github.com/NVIDIA/attestation-sdk", tag = "2026.03.02", optional = true }
+
 [[bin]]
 name = "evidence_getter"
 required-features = ["bin"]
@@ -76,6 +78,7 @@ all-attesters = [
     "cca-attester",
     "se-attester",
     "tpm-attester",
+    "nvidia-attester",
 ]
 
 # tsm-report enables a module that helps attesters to use Linux TSM_REPORTS for generating
